+++ tests/testthat/test-validation-same-length.R
@@ -34,5 +34,8 @@ test_that("validateIsSameLength returns `NULL` when objects have same lengths", 
 
 test_that("validateIsSameLength produces errors when objects do not have same lengths", {
   errorMessageIsSameLength <- "Arguments 'A, B' must have the same length, but they don't!"
-  expect_error(validateIsSameLength(A, B), errorMessageIsSameLength)
+  expect_error(
+    validateIsSameLength(A, B), 
+    "(Arguments).*(A, B).*(must have the same length, but they don't)"
+    )
 })