+++ NEWS.md
@@ -1,5 +1,12 @@
 # ospsuite.utils (development version)
 
+## Major changes
+
+- New functions using `{cli}` and `{logger}` packages to record and keep tracks of what the code does in a log file:
+  - `setLogFolder()`: initialize/end the logging system
+  - `logCatch()`: catch and record messages, warnings, and errors
+  - `logError()`, `logWarning()`, `logInfo()`, `logDebug()`: record errors, warnings, info and debug messages in the logs.
+
 # ospsuite.utils 1.7.0
 
 ## Major changes